+++ contracts/game-of-chance/CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.1.1] - 2022-10-26
+### Changed
+- Updated fungible token to v0.1.2.
+- Updated `gstd`, `gtest`, `gear-wasm-builder` to the current commit hash (`#d4552434`).
 
 ## [0.1.0] - 2022-07-27
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/gear-dapps/lottery/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/gear-dapps/lottery/compare/0.1.1...HEAD
+[0.1.1]: https://github.com/gear-dapps/lottery/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/gear-dapps/lottery/compare/60d5a8e...0.1.0